# documentation/api/OPENAPI-3.0.yaml
openapi: 3.0.0
info:
  title: Sugar-Daddy Platform - Comprehensive API
  description: |
    完整的 Sugar-Daddy 平台 API 文檔
    包含所有 81 個端點的詳細說明
    
    ### 系統組成
    - 🔐 Auth Service (11 個端點)
    - 🎬 Content-Streaming Service (11 個端點)  
    - 💳 Payment Service (19 個端點)
    - 🎁 Subscription Service (10 個端點)
    - 🤖 Recommendation Service (13 個端點)
    - 🌉 API Gateway (1 個端點)
    
    **總計: 81 個 API 端點**
  version: 1.0.0
  contact:
    name: Sugar-Daddy Backend Team
    email: backend@sugardaddy.com
  license:
    name: Proprietary
    url: https://sugardaddy.com/license
servers:
  - url: http://localhost:3000
    description: API Gateway (本地開發)
    variables:
      port:
        default: '3000'
  - url: https://api.sugardaddy.com/v1
    description: 生產環境
  - url: https://staging-api.sugardaddy.com/v1
    description: 預發佈環境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token 認證
        登錄後獲得 token，在 Authorization header 中使用
        格式: Authorization: Bearer <token>
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: Bad Request
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required:
        - statusCode
        - message

    # ==================== Auth Service Schemas ====================
...
    UploadSession:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        filename:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        uploadedBytes:
          type: integer
        expiresAt:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: 未授權
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 401
            message: Unauthorized
            errors:
              - field: token
                message: Invalid or expired token
    NotFoundError:
      description: 資源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            statusCode: 404
            message: Not Found
            errors:
              - message: Resource not found
    BadRequestError:
      description: 請求格式錯誤
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
paths:
...
  /api/uploads/initiate:
    post:
      tags:
        - Content-Streaming
      summary: 初始化上傳會話
      description: 開始分片上傳，獲得會話 ID
      operationId: initiateUpload
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
                  example: my-video.mp4
                contentType:
                  type: string
                  example: video/mp4
                fileSize:
                  type: integer
                  example: 524288000
              required:
                - filename
                - contentType
                - fileSize
      responses:
        '201':
          description: 上傳會話已創建
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    $ref: '#/components/schemas/UploadSession'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/uploads/{sessionId}:
    get:
      tags:
        - Content-Streaming
      summary: 獲取上傳狀態
      description: 獲取指定上傳會話的當前狀態
      operationId: getUploadStatus
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 上傳狀態
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    $ref: '#/components/schemas/UploadSession'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/uploads/{sessionId}/chunk:
    post:
      tags:
        - Content-Streaming
      summary: 上傳分片
      description: 上傳視頻文件的一部分
      operationId: uploadChunk
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: chunkIndex
          in: query
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '200':
          description: 分片上傳成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: object
                    properties:
                      chunkIndex:
                        type: integer
                      uploadedBytes:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/uploads/{sessionId}/complete:
    post:
      tags:
        - Content-Streaming
      summary: 完成上傳
      description: 完成上傳會話，開始轉碼
      operationId: completeUpload
      security:
        - BearerAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
              required:
                - title
      responses:
        '200':
          description: 上傳完成，轉碼已啟動
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                  data:
                    type: object
                    properties:
                      videoId:
                        type: string
                        format: uuid
                      message:
                        type: string
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /api/recommendations/{contentId}/view:
    post:
      tags:
        - Recommendation
      summary: 記錄觀看
      description: 記錄用戶對內容的觀看
      operationId: recordView
      security:
        - BearerAuth: []
      parameters:
        - name: contentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                watchDuration:
                  type: integer
                  description: 觀看時長（秒）
              required:
                - watchDuration
      responses:
        '200':
          description: 觀看已記錄
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
security:
  - BearerAuth: []